refactor(Utils): move geo stuff to a dedicated file
